Declare Minter in DatabaseManager
First step to make the code more modular. As of right now, the
MinterController directly depends on both the DatabaseManager and the
Minter; the latter of which also depends on the former. My intentions
are to make the Minter completely independent and become a 'dumb' id
generator. In addition I also want to make the DatabaseManager to be the
middle ground between the MinterController and the Minter.

changes made to this commit:
- declared a Minter as a field in DatabaseManager
- created methods to instantiate an appropriate Minter to use

missing changes to be made:
- add missing javadoc