Refactor the names of the files
The idea behind this is that by simplifying the Minter, now
IdGenerators, it no longer needs an instance of the
DatabaseManager as its sole function is to produce a set of ids.
By definition, the act of minting authorize an object for use.
As far as the service is concerned, an id is officially in use
whenever it is entered into the database.

missing changes:
- javadocs still need to be added to new methods
- certain javadocs need to be adjusted since the renaming