Correct rootLength test cases
An error was occured when testing rootLengths. Previously the
generators would, by default, generate 1000 ids, regardless of whether
or not it was possible for them to do so. This is now corrected. Now
whenever a specific combination of parameters can only less than 1000
ids, the generators will produce the maximum possible value instead.