Separate testGetCurrentSetting into two methods
When using the getDefaultSetting method there is an internal check to
see if the DefaultSetting table contains a single entry:

	If there isn't an entry then one is hard-written into the table.
	This situtation should only occur when the application creates a
	new database. After the entry has been created the method then
	returns it.

	Otherwise, the application returns the entry.

For 100% code coverage, the alternative branch must be satisfied by
dissolving the original	method into two separate methods. One tests
the scenario where the entry already exists and	the other tests the
scenario where it doesn't.

additional changes:
- remove SQLException from javadoc in mint method in MinterServiceImpl
- rename test method updateCurrentSetting to testUpdateCurrentSetting
- create global variable DefaultSetting and have it be initialized in
     	the method initializeDefaultSetting.