Correct undefined 'BOOLEAN' datatypes for DbUnit
There was a bug with DbUnit that prevented boolean types in an XML
dataset from being read into a HSQL database, causing a
NoSuchColumnException. This is because HSQL uses their own definition
of Boolean which is not understood by DbUnit's DefaultDataTypeFactory.

This issue was remedied by adding a HsqlDataTypeFactory that would
replace HSQL's boolean with DbUnit's boolean. HsqlDataTypeFactory is
during testing called whenever a connection is being made to the
DataSource.