Clean ResolverService and ResolverController
ResolverService:
All methods with the return type of boolean were changed to void.
After adding the generalExceptionHandler method in the
ResolverController, there was no need to return null. Combine with the
fact that whenever Hibernate fails to perform a transaction, it throws
its own exceptions, the blocks of code that would send "null" to the
view resolver were unreachable. Some test cases had to be re-written
so that they could be properly mocked by Mockito.

In an effort to clean the following changes were made to
ResolverController:
- implemented the changes noted above, removing if/else blocks
- added in-line comments
- updated javadocs
- renamed convertToJsonObject to convertPurlToJson because json
	strings aren't objects, they're just a notation that contain
	objects