Refactor Purl and its fields
A conclusion was made that the POJO "Purl" was a unity of transaction
where logical operations could be made. An actual "PURL" is a
persistent URL that is designed to be unique across all domains
(hence the word persistent). As such, the class "Purl" was renamed
to "Citation" to better match this description. In addition, the
tentative field, "Pid" was renamed to "Purl" to uniquely identify
a Citaiton object.Several local fields and javadocs were renamed to
match this change.