Integrate SpringBoot into PidDaoTest
Since Spring boot is being used there was a hidden error that prevented
the Dao tests from running. This is because TestNg and Dbunit were
creating its own application context, interfering with Spring Boot's.
As a result, the respository objects are instantiated using the
setter methods created by the class RepositoryConfiguration.

The following changes were made:
- SpringApplicationConfiguration annotation was added
- removal of unneeded libraries
- non-unique Pids no longer throws exceptions, count is used instead
- imports strictly used by Dbunit are removed
- the class no longer extends EntityDaoImplTest
- Pid.xml is deleted