Add methods to read and from properties file
MinterService has been given methods to read and write from
the properties file. The storage in the properties file will be used
alongside its storage in Hibernate, not replacing it. For now, these
methods will be called whenever getCurrentSetting and
updateCurrentSetting is called.

The methods were given a string parameter called 'filePath' to
interact with the properties file. This is done to make testing easier
and gives users the option of explicitly defining where the properties
file is located. For now, the value of the path is stored as a constant
in the MinterController class.

The tests were updated to reflect this change. Reading and writing
is not directly tested at the moment.