Assign characters to enum TokenType
Characters to each token to remove the need for an external source
to associate values to them. Now each TokenType references a String of
characters that can be called using the getCharacters method.

The following fields were added to express characters without vowels:
- LOWER_CONSONANTS
- UPPER_CONSONANTS
- MIXED_CONSONANTS
- LOWER_CONSONANTS_EXTENDED
- UPPER_CONSONANTS_EXTENDED
- MIXED_CONSONANTS_EXTENDED

The following fields were renamed to better reflect the difference
between the fields with and without consonants:
- "LOWERCASE" was renamed to "LOWER_ALPHABET"
- "UPPERCASE" was renamed to "UPPER_ALPHABET"
- "MIXEDCASE" was renamed to "MIXED_ALPHABET"
- "LOWER_EXTENDED" was renamed to "LOWER_ALPHABET_EXTENDED"
- "UPPER_EXTENDED" was renamed to "UPPER_ALPHABET_EXTENDED"
- "MIXED_EXTENDED" was renamed to "MIXED_ALPHABET_EXTENDED"