Use getResourceAsStream to read/write properties
The minter won't be able to find the configuration file in a real
deployment since the Minter.jar could be in any directory. Since Maven
will put anything under src/main/resources in the JAR though and it can
be loaded using getResourceAsStream() instead of a normal file open.
As a result the paths to all properties files have been changed so it's
relative to the resources folders.