Remove JUnit calls from repository test classes
In order to use TestNG an application context must be provided and the
classes must extend AbstractTestNGSpringContextTests. A full
application context is provided by and set-up using the
WebAppConfiguration and the Integration test annotations. However,
extending AbstractTestNGSpringContextTests enables
ServletTestExecutionListener, asserting that the newly created
application context is a mock. To correct this, we must add the
TestExecutionListeners annotation to disable its inherited listeners.

additional changes:
- replaced JUnit's After annotation for TestNG's AfterMethod
- auto-formatted code
- added AutoWired annotations directly to repository classes
- removed set methods used by JUnit's class runner
- updated class javadocs
- updated test cases in UsedSettingRepositoryTest