Merge functionality of test*.properties files
These two files were created to separately test the reading and writing
functionality of the MinterService's updateCurrentSetting method. Now
that the properties library is in use to modify the project's
generated properties file, their separation is no longer needed as
there isn't a risk of accidentally overriding the test files.

additional changes:
- remove unused import in MinterService
- added method to restore changes made to testDefaultSetting during
	MinterService testing