Correct longToName methods
When attempting to create Pids with using settings, the maximum
permutation can well exceed that of an integer, creating a wrap-around
if a long number is cast into a digit. The lines of code that were
corrected were prematurely casting values into ints, leading to
negative numbers.

side-note:
Given some

a mod b = c

Mathematically, if b is positive then the result c will be c >= 0.
This result is defined by the modulus operation. In Java however, the
'%' symbol is modulus but rather the remainder. If a < 0, then c <= 0.
This result, as well as numbers wrapping around should they exceed
the limit of an int, explains why negative numbers were produced.