Replace TreeSet with LinkedHashSet
TreeSet is replaced for two reasons. The first is that although the
MinterController returned the order in which the Pids were added to
a set, the natural order (defined by compareTo) was used everywhere
else. This was fine up until sequentialMint needed to be tested with
an arbitrary starting value, making a Pid's order of membership in a
set preffered over their sorted, natural order. The second reason is
that the basic operations (add, remove, delete) of LinkedHashSet are
faster.

additional changes:
The equals() and hashCode() needed to be overridden in order for
LinkedHashSet to work properly.