Add a text field for window cache
This is a text field where the user can adjust the size of their cache.
A cache is a temporary collection of Pids that is created whenever the
application is re/started. The format of the Pids are determined by
the default settings. As Pids are created using the default values,
they're drawn from the cache.

additional changes:
- remove unused lines of code
- remove submitForm()