Initialize StoredSetting during context refresh
As it stands there is no reason to have the StoredSetting field only
be initialized whenever getStoredSetting is called, a method that has
an unnecessary dual purpose. As a result, getStoredSetting has been
separated into two methods. The first is to initialize StoredSetting
and the other is to return its value. Alongside the cache, the
StoredSetting field will now be initialized whenever the context is
refreshed.