Add methods to setup MinterService class
During deployment, the StoredSetting field will  be initialized
whenever the context is refreshed or restarted. Since an application
context is not created, the field is set up manually instead.

In addition, because we're testing the MinterService using a test
properties file, the DefaultSettingPath field needs to be changed.

additional changes:
- add a cahceSize key